support two types of Intel GVT-g compositor display
There are two types of Intel GVT-g compositor display.
One is using egl: we generate a texture based on the guest fb and
bind the texture to the QEMU window. This type of compositor display
only can work with SDL and does not support remove display.

The other one is we copy the guest fb to QEMU and QEMU will display
guest fb automatically. This type of compositor display can work with
both SDL and remote display like vnc and spice. We use blt copy to
copy the guest fb to userspace(QEMU) to accelerate the copy process.
Because there is a copy so the performance is a little drop compared
to the first one.

Following is our policy:
1. default is the first one(calling egl implementation).
2. if the egl implementation is not available(no SDL or egl extention)
   then the second one(calling blt implementation) will be enabled.
3. if egl/blt are both enabled only egl implementation will be enabled.
4. if you want to use blt implementation disable egl implementation while
   config.

Signed-off-by: xiaoguang chen <xiaoguang.chen@intel.com>