fix linux desktop can not diplay bug
When loging into console mode the fb's tiling is linear(not tiled) if we
start the lighdm service the fb is tiled but sometimes the fb's address(
from vgtbuffer ioCtl) did not change while the tiling changed.
Based on current logic we only compare the fb's start address if the
address is the same we will reuse the vgtbuffer which will lead to fb's
display wrong if the fb's start address did not change but the tiling
changed.
The tiling changes implies two issues:
1. The framebuffer's stride will change. Since "eglCreateImageKHR()"
needs the stride as input attribute, we need to recreate eglImage.
2. The backing storage of framebuffer may change. In taht case, we need
to update the backing stroage of vgtbuffer as well.
Since the graphics memory address of the framebuffer does not change,
the compositormissed the above two changes.
This patch could address issue 1 completely and issue 2 partly. We
could not know if the GTT entry changed for the backing storage. There
must be some kind of notification mechanism to inform us of this.

The complete fix of issue 2 needs some checking and notification from
vgt device model. Once we implement the notification in vgt I will
complete the fix to issue 2.

Now we will compare the start address/tiling/size of a fb to decide
whether reuse the vgtbuffer or not.

Signed-off-by: xiaoguang chen <xiaoguang.chen@intel.com>