Handle correctly when guest frambuffer is changed
Bug 1181 - qemu window displays blurred after starting X and switching to dom0

To reduce flicker of desktop, i915 driver maintain multiple start address for guest framebuffer.
A corresponding list is maintained in QEMU to record the start address, it can reduce ioctls if the start address is switched among the list items.

When lightdm in guest VM is started and stopped in order, the framebuffer of guest VM will be created and destroyed.
When the framebuffer is re-created, it may be allocated with the same address, but the buffer is destroyed in the previous operation.
The framebuffer should be re-binded as texture, re-use the previous one is not a correct choice.

There is no mechnism that QEMU is notified when the framebuffer is destroyed, so the variable "tiled" and "size" is used to help to distinguish the situation.
As the "tiled" and "size" in the list is assumed to be the same for one guest VM, so if the variable changes, it is assumed that the framebuffer should be re-binded.

Signed-off-by: Zhao, Xinda <xinda.zhao@intel.com>
Reviewed-by: Zheng, Xiao <xiao.zheng@intel.com>